Move registration and login form to components
